+++ crates/forge/tests/integration/get_available_gas.rs
@@ -43,7 +43,7 @@ fn test_get_available_gas() {
         "#
         ),
         Contract::from_code_path(
-            "HelloStarknet".to_string(),
+            "contract::HelloStarknet".to_string(),
             Path::new("tests/data/contracts/hello_starknet.cairo"),
         )
         .unwrap()