+++ crates/forge/tests/integration/get_class_hash.rs
@@ -24,7 +24,7 @@ fn get_class_hash_cheatcode() {
         "#
         ),
         Contract::from_code_path(
-            "HelloStarknet".to_string(),
+            "contract::HelloStarknet".to_string(),
             Path::new("tests/data/contracts/hello_starknet.cairo"),
         )
         .unwrap()