+++ crates/sncast/src/helpers/scarb_utils.rs
@@ -1,17 +1,16 @@
-use crate::helpers::artifacts::CastStarknetContractArtifacts;
+use crate::helpers::artifacts::{CastStarknetContractArtifacts, ContractArtifactsMap};
 use anyhow::{Context, Result, anyhow};
 use camino::{Utf8Path, Utf8PathBuf};
 use foundry_ui::{UI, components::warning::WarningMessage};
 use scarb_api::metadata::{MetadataError, MetadataOpts, metadata_for_dir, metadata_with_opts};
 use scarb_api::{
-    CompilationOpts, ScarbCommand, ScarbCommandError, ensure_scarb_available,
+    CompilationOpts, ContractsData, ScarbCommand, ScarbCommandError, ensure_scarb_available,
     get_contracts_artifacts_and_source_sierra_paths,
     metadata::{Metadata, PackageMetadata},
     target_dir_for_workspace,
 };
 use scarb_ui::args::PackagesFilter;
 use shared::command::CommandExt;
-use std::collections::HashMap;
 use std::str::FromStr;
 
 pub fn get_scarb_manifest() -> Result<Utf8PathBuf> {
@@ -142,12 +141,28 @@ pub fn build(
     cmd.run()
 }
 
+fn contracts_data_to_artifacts(contracts: ContractsData) -> ContractArtifactsMap {
+    contracts
+        .into_iter()
+        .map(|(module_path, contract)| {
+            (
+                module_path,
+                CastStarknetContractArtifacts {
+                    sierra: contract.artifacts.sierra,
+                    casm: serde_json::to_string(&contract.artifacts.casm)
+                        .expect("serialization on valid Casm cannot fail"),
+                },
+            )
+        })
+        .collect()
+}
+
 pub fn build_and_load_artifacts(
     package: &PackageMetadata,
     config: &BuildConfig,
     build_for_script: bool,
     ui: &UI,
-) -> Result<HashMap<String, CastStarknetContractArtifacts>> {
+) -> Result<ContractArtifactsMap> {
     // TODO (#2042): Remove this logic, always use release as default
     let default_profile = if build_for_script { "dev" } else { "release" };
     build(package, config, default_profile)
@@ -163,10 +178,8 @@ pub fn build_and_load_artifacts(
             ui,
             CompilationOpts::default(),
         )
-        .context("Failed to load artifacts. Make sure you have enabled sierra code generation in Scarb.toml")?
-        .into_iter()
-        .map(|(name, (artifacts, _))| (name, CastStarknetContractArtifacts { sierra: artifacts.sierra, casm: serde_json::to_string(&artifacts.casm).expect("valid serialization") }))
-        .collect())
+        .context("Failed to load artifacts. Make sure you have enabled sierra code generation in Scarb.toml")
+        .map(contracts_data_to_artifacts)?)
     } else {
         let profile = &config.profile;
         ui.println(&WarningMessage::new(&format!(
@@ -178,10 +191,8 @@ pub fn build_and_load_artifacts(
             ui,
             CompilationOpts::default(),
         )
-        .context("Failed to load artifacts. Make sure you have enabled sierra code generation in Scarb.toml")?
-        .into_iter()
-        .map(|(name, (artifacts, _))| (name, CastStarknetContractArtifacts { sierra: artifacts.sierra, casm: serde_json::to_string(&artifacts.casm).expect("valid serialization") }))
-        .collect())
+        .context("Failed to load artifacts. Make sure you have enabled sierra code generation in Scarb.toml")
+        .map(contracts_data_to_artifacts)?)
     }
 }