+++ crates/sncast/tests/e2e/main_tests.rs
@@ -3,7 +3,8 @@ use crate::helpers::constants::{
 };
 use crate::helpers::env::set_keystore_password_env;
 use crate::helpers::fixtures::{
-    copy_file, duplicate_contract_directory_with_salt, get_accounts_path, get_keystores_path,
+    copy_directory_to_tempdir, copy_file, duplicate_contract_directory_with_salt,
+    get_accounts_path, get_keystores_path,
 };
 use crate::helpers::runner::runner;
 use configuration::test_utils::copy_config_to_tempdir;
@@ -140,7 +141,7 @@ async fn test_happy_case_from_cli_with_sncast_config() {
 
     snapbox.assert().success().stdout_eq(indoc! {r"
         Success: Call completed
-        
+
         Response:     0x0
         Response Raw: [0x0]
     "});
@@ -169,7 +170,7 @@ async fn test_happy_case_mixed() {
 
     snapbox.assert().success().stdout_eq(indoc! {r"
         Success: Call completed
-        
+
         Response:     0x0
         Response Raw: [0x0]
     "});
@@ -347,19 +348,11 @@ async fn test_keystore_inexistent_account() {
 async fn test_keystore_undeployed_account() {
     let contract_path =
         duplicate_contract_directory_with_salt(CONTRACTS_DIR.to_string() + "/map", "put", "8");
-    // TODO(#4311): Remove temporary `--accounts-file` workaround for `devnet-<i>`.
-    let accounts_file = "empty_accounts.json";
-    copy_file(
-        "tests/data/accounts/empty_accounts.json",
-        contract_path.path().join(accounts_file),
-    );
     let my_key_path = get_keystores_path("tests/data/keystore/my_key.json");
     let my_account_undeployed_path =
         get_keystores_path("tests/data/keystore/my_account_undeployed.json");
 
     let args = vec![
-        "--accounts-file",
-        accounts_file,
         "--keystore",
         my_key_path.as_str(),
         "--account",
@@ -434,4 +427,24 @@ async fn test_keystore_and_ledger_conflict() {
     let output = snapbox.assert().failure();
 
     assert_stderr_contains(output, "Error: keystore and ledger cannot be used together");
+}
+
+#[test]
+fn test_accepts_full_module_path_for_ambiguous_contract_name() {
+    let contract_path =
+        copy_directory_to_tempdir(CONTRACTS_DIR.to_string() + "/duplicate_contract_name");
+
+    let args = vec![
+        "utils",
+        "class-hash",
+        "--contract-name",
+        "duplicate_contract_name::first_contract::HelloStarknet",
+    ];
+
+    let output = runner(&args)
+        .current_dir(contract_path.path())
+        .assert()
+        .success();
+
+    assert_stdout_contains(output, indoc! {r"Class Hash: 0x0[..]"});
 }